move getworker use case code into worker usecase
