+++ SocialCareCaseViewerApi/V1/UseCase/Interfaces/IWorkersUseCase.cs
@@ -1,3 +1,4 @@
+using System.Collections.Generic;
 using SocialCareCaseViewerApi.V1.Boundary.Requests;
 using SocialCareCaseViewerApi.V1.Boundary.Response;
 
@@ -5,6 +6,8 @@ namespace SocialCareCaseViewerApi.V1.UseCase.Interfaces
 {
     public interface IWorkersUseCase
     {
+        List<WorkerResponse> ExecuteGet(GetWorkersRequest request);
+
         WorkerResponse ExecutePost(CreateWorkerRequest createWorkerRequest);
 
         void ExecutePatch(UpdateWorkerRequest updateWorkerRequest);
